refactor(actions): Decompose large action file for better organization
